A merge with nothing to merge is not handed out
A merge atom pins its sixteen children when the DAG is built (Invariant 2),
with '' for one that was not published then. All sixteen empty means it can
only fail — merge-v1 throws "no published child of z/x/y" — and since it is
ready with no deps, every tab claims it, fails, and claims it again. A world
compiles from the leaves up, so on a new world that loop is all a tab sees:

    claim 8 merge
    error 8 merge — run@…/atoms/merge.js:223

claim_atom now skips those. Nothing has to rewrite them: publishing a child
marks the parent dirty (0006), which opens a job whose DAG pins the child.

Fixtures that asserted the old behaviour now publish one child first
(db/test/0005, 0006, 0011, client/test/e2e/background).

Three GIS tests were failing before this change and are fixed here, since the
gate has to be green to mean anything: st_ndims() is a smallint and pgTAP's
is() would not compare it with an int literal (0029, 0031, 0034), and 0031
still asserted gis.feature, which 0034 deliberately replaced with one typed
layer per kind.

make db-test: 29 files, 476 tests, green.

+++ db/test/0029_gistables.sql
@@ -22,7 +22,7 @@ INSERT INTO feature (kind, geom)
 VALUES ('road', st_geomfromtext('LINESTRING(7.01 46.01, 7.02 46.02)', 4326));
 
 SELECT is(
-    (SELECT st_ndims(geom) FROM feature WHERE kind = 'road' LIMIT 1),
+    (SELECT st_ndims(geom)::int FROM feature WHERE kind = 'road' LIMIT 1),
     3,
     'a 2D line drawn in QGIS is stored with a Z'
 );
